fix(task-20260704-4): resolve admin dashboard and event detail activity formatting crashes
files: app/src/main/java/vn/humg/hai/event_ticket_booking_app/fragments/AdminDashboardFragment.java, app/src/main/java/vn/humg/hai/event_ticket_booking_app/view/EventDetailActivity.java, app/src/main/java/vn/humg/hai/event_ticket_booking_app/adapter/TransactionAdapter.java
review: 1 rounds
note: " Cast double prices to long to prevent IllegalFormatConversionException and added null checks for optional reporting UI elements on admin dashboard.\

+++ app/src/main/java/vn/humg/hai/event_ticket_booking_app/adapter/TransactionAdapter.java
@@ -41,7 +41,7 @@ public class TransactionAdapter extends RecyclerView.Adapter<TransactionAdapter.
         // 1. Hiển thị Mã đơn và Giá tiền (Sử dụng string resource)
         holder.tvId.setText(String.format(Locale.getDefault(), "#%s", booking.getBookingId().toUpperCase()));
         
-        String priceText = holder.itemView.getContext().getString(R.string.price_format, booking.getTotalPrice());
+        String priceText = holder.itemView.getContext().getString(R.string.price_format, (long) booking.getTotalPrice());
         holder.tvPrice.setText(priceText);
         
         // 2. Hiển thị số lượng vé
